Refactor, add TwoDimensionalTableUtil
